Ported over updates from recently-merged class.
